requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
chromadb
streamlit
python-dotenv
pypdf
pandas
google-generativeai
openpyxl